move EvolutionItem to its own file
