+++ src/enums/evolution-item.ts
@@ -1,0 +1,48 @@
+export enum EvolutionItem {
+  NONE,
+
+  LINKING_CORD,
+  SUN_STONE,
+  MOON_STONE,
+  LEAF_STONE,
+  FIRE_STONE,
+  WATER_STONE,
+  THUNDER_STONE,
+  ICE_STONE,
+  DUSK_STONE,
+  DAWN_STONE,
+  SHINY_STONE,
+  CRACKED_POT,
+  SWEET_APPLE,
+  TART_APPLE,
+  STRAWBERRY_SWEET,
+  UNREMARKABLE_TEACUP,
+  UPGRADE,
+  DUBIOUS_DISC,
+  DRAGON_SCALE,
+  PRISM_SCALE,
+  RAZOR_CLAW,
+  RAZOR_FANG,
+  REAPER_CLOTH,
+  ELECTIRIZER,
+  MAGMARIZER,
+  PROTECTOR,
+  SACHET,
+  WHIPPED_DREAM,
+  SYRUPY_APPLE,
+  CHIPPED_POT,
+  GALARICA_CUFF,
+  GALARICA_WREATH,
+  AUSPICIOUS_ARMOR,
+  MALICIOUS_ARMOR,
+  MASTERPIECE_TEACUP,
+  SUN_FLUTE,
+  MOON_FLUTE,
+
+  BLACK_AUGURITE = 51,
+  PEAT_BLOCK,
+  METAL_ALLOY,
+  SCROLL_OF_DARKNESS,
+  SCROLL_OF_WATERS,
+  LEADERS_CREST
+}
