+++ src/enums/field-position.ts
@@ -1,4 +1,3 @@
-
 export enum FieldPosition {
   CENTER,
   LEFT,
